cambios finales 2.0

+++ routes/web.php
@@ -29,13 +29,13 @@ Auth::routes();
 Route::get('/', [App\Http\Controllers\ColeccioneController::class, 'listaColecciones'])->name('home');
 
 //ADMIN
-Route::get('/colecciones/create', [ColeccioneController::class,'crear'])->name('colecciones.crear');
-Route::get('/productos/create/{id}', [ProductoController::class,'crear'])->name('productos.crear');
-Route::get('/tallas/create/{id}', [TallaController::class,'crear'])->name('tallas.crear');
-Route::get('/colecciones/{coleccion}/edit', [ColeccioneController::class,'editar'])->name('colecciones.editar');
-Route::get('/productos/{producto}/edit', [ProductoController::class,'editar'])->name('productos.editar');
-Route::get('/ventas', [PedidoController::class,'mostrarVentas'])->name('pedidos.mostrarVentas');
-Route::get('/listado-usuarios', [UserController::class,'listadoUsuarios'])->name('users.listadoUsuarios');
+Route::get('/colecciones/create', [ColeccioneController::class,'crear'])->name('colecciones.crear')->middleware(['role:admin']);
+Route::get('/productos/create/{id}', [ProductoController::class,'crear'])->name('productos.crear')->middleware(['role:admin']);
+Route::get('/tallas/create/{id}', [TallaController::class,'crear'])->name('tallas.crear')->middleware(['role:admin']);
+Route::get('/colecciones/{coleccion}/edit', [ColeccioneController::class,'editar'])->name('colecciones.editar')->middleware(['role:admin']);
+Route::get('/productos/{producto}/edit', [ProductoController::class,'editar'])->name('productos.editar')->middleware(['role:admin']);
+Route::get('/ventas', [PedidoController::class,'mostrarVentas'])->name('pedidos.mostrarVentas')->middleware(['role:admin']);
+Route::get('/listado-usuarios', [UserController::class,'listadoUsuarios'])->name('users.listadoUsuarios')->middleware(['role:admin']);
 
 //COLECCIONES-PRODUCTOS
 Route::get('/colecciones/{coleccion}', [ProductoController::class,'listaProductos'])->name('productos.listaProductos');
@@ -68,21 +68,21 @@ Route::view('/aviso-legal', 'footerInfo.avisoLegal')->name('footerInfo.avisoLega
 
 
 //CRUD COLECCIONES
-Route::get('panel/colecciones', [ColeccioneController::class,'index'])->name('colecciones.index');
-Route::get('panel/colecciones/create', [ColeccioneController::class,'create'])->name('colecciones.create');
-Route::get('panel/colecciones/{coleccion}', [ColeccioneController::class,'show'])->name('colecciones.show');
-Route::get('panel/colecciones/{coleccion}/edit', [ColeccioneController::class,'edit'])->name('colecciones.edit');
-Route::post('panel/colecciones', [ColeccioneController::class,'store'])->name('colecciones.store');
-Route::patch('panel/colecciones/{coleccion}', [ColeccioneController::class,'update'])->name('colecciones.update');
-Route::delete('panel/colecciones/{coleccion}/destroy', [ColeccioneController::class,'destroy'])->name('colecciones.destroy');
+Route::get('panel/colecciones', [ColeccioneController::class,'index'])->name('colecciones.index')->middleware(['role:admin']);
+Route::get('panel/colecciones/create', [ColeccioneController::class,'create'])->name('colecciones.create')->middleware(['role:admin']);
+Route::get('panel/colecciones/{coleccion}', [ColeccioneController::class,'show'])->name('colecciones.show')->middleware(['role:admin']);
+Route::get('panel/colecciones/{coleccion}/edit', [ColeccioneController::class,'edit'])->name('colecciones.edit')->middleware(['role:admin']);
+Route::post('panel/colecciones', [ColeccioneController::class,'store'])->name('colecciones.store')->middleware(['role:admin']);
+Route::patch('panel/colecciones/{coleccion}', [ColeccioneController::class,'update'])->name('colecciones.update')->middleware(['role:admin']);
+Route::delete('panel/colecciones/{coleccion}/destroy', [ColeccioneController::class,'destroy'])->name('colecciones.destroy')->middleware(['role:admin']);
 //CRUD PRODUCTOS
-Route::get('panel/productos', [ProductoController::class,'index'])->name('productos.index');
-Route::get('panel/productos/create', [ProductoController::class,'create'])->name('productos.create');
-Route::get('panel/productos/{producto}', [ProductoController::class,'show'])->name('productos.show');
-Route::get('panel/productos/{producto}/edit', [ProductoController::class,'edit'])->name('productos.edit');
-Route::post('panel/productos', [ProductoController::class,'store'])->name('productos.store');
-Route::patch('panel/productos/{producto}', [ProductoController::class,'update'])->name('productos.update');
-Route::delete('panel/productos/{producto}/destroy', [ProductoController::class,'destroy'])->name('productos.destroy');
+Route::get('panel/productos', [ProductoController::class,'index'])->name('productos.index')->middleware(['role:admin']);
+Route::get('panel/productos/create', [ProductoController::class,'create'])->name('productos.create')->middleware(['role:admin']);
+Route::get('panel/productos/{producto}', [ProductoController::class,'show'])->name('productos.show')->middleware(['role:admin']);
+Route::get('panel/productos/{producto}/edit', [ProductoController::class,'edit'])->name('productos.edit')->middleware(['role:admin']);
+Route::post('panel/productos', [ProductoController::class,'store'])->name('productos.store')->middleware(['role:admin']);
+Route::patch('panel/productos/{producto}', [ProductoController::class,'update'])->name('productos.update')->middleware(['role:admin']);
+Route::delete('panel/productos/{producto}/destroy', [ProductoController::class,'destroy'])->name('productos.destroy')->middleware(['role:admin']);
 //CRUD COMENTARIOS
 Route::get('panel/comentarios', [ComentarioController::class,'index'])->name('comentarios.index');
 Route::get('panel/comentarios/create', [ComentarioController::class,'create'])->name('comentarios.create');
@@ -92,21 +92,21 @@ Route::post('panel/comentarios', [ComentarioController::class,'store'])->name('c
 Route::patch('panel/comentarios/{comentario}', [ComentarioController::class,'update'])->name('comentarios.update');
 Route::delete('panel/comentarios/{comentario}/destroy', [ComentarioController::class,'destroy'])->name('comentarios.destroy');
 //CRUD PEDIDOS
-Route::get('panel/pedidos', [PedidoController::class,'index'])->name('pedidos.index');
-Route::get('panel/pedidos/create', [PedidoController::class,'create'])->name('pedidos.create');
-Route::get('panel/pedidos/{pedido}', [PedidoController::class,'show'])->name('pedidos.show');
-Route::get('panel/pedidos/{pedido}/edit', [PedidoController::class,'edit'])->name('pedidos.edit');
-Route::post('panel/pedidos', [PedidoController::class,'store'])->name('pedidos.store');
-Route::patch('panel/pedidos/{pedido}', [PedidoController::class,'update'])->name('pedidos.update');
-Route::delete('panel/pedidos/{pedido}/destroy', [PedidoController::class,'destroy'])->name('pedidos.destroy');
+Route::get('panel/pedidos', [PedidoController::class,'index'])->name('pedidos.index')->middleware(['role:admin']);
+Route::get('panel/pedidos/create', [PedidoController::class,'create'])->name('pedidos.create')->middleware(['role:admin']);
+Route::get('panel/pedidos/{pedido}', [PedidoController::class,'show'])->name('pedidos.show')->middleware(['role:admin']);
+Route::get('panel/pedidos/{pedido}/edit', [PedidoController::class,'edit'])->name('pedidos.edit')->middleware(['role:admin']);
+Route::post('panel/pedidos', [PedidoController::class,'store'])->name('pedidos.store')->middleware(['role:admin']);
+Route::patch('panel/pedidos/{pedido}', [PedidoController::class,'update'])->name('pedidos.update')->middleware(['role:admin']);
+Route::delete('panel/pedidos/{pedido}/destroy', [PedidoController::class,'destroy'])->name('pedidos.destroy')->middleware(['role:admin']);
 //CRUD UNIONPEDIDOS
-Route::get('panel/unionpedidos', [UnionpedidoController::class,'index'])->name('unionpedidos.index');
-Route::get('panel/unionpedidos/create', [UnionpedidoController::class,'create'])->name('unionpedidos.create');
-Route::get('panel/unionpedidos/{unionpedido}', [UnionpedidoController::class,'show'])->name('unionpedidos.show');
-Route::get('panel/unionpedidos/{unionpedido}/edit', [UnionpedidoController::class,'edit'])->name('unionpedidos.edit');
-Route::post('panel/unionpedidos', [UnionpedidoController::class,'store'])->name('unionpedidos.store');
-Route::patch('panel/unionpedidos/{unionpedido}', [UnionpedidoController::class,'update'])->name('unionpedidos.update');
-Route::delete('panel/unionpedidos/{unionpedido}/destroy', [UnionpedidoController::class,'destroy'])->name('unionpedidos.destroy');
+Route::get('panel/unionpedidos', [UnionpedidoController::class,'index'])->name('unionpedidos.index')->middleware(['role:admin']);
+Route::get('panel/unionpedidos/create', [UnionpedidoController::class,'create'])->name('unionpedidos.create')->middleware(['role:admin']);
+Route::get('panel/unionpedidos/{unionpedido}', [UnionpedidoController::class,'show'])->name('unionpedidos.show')->middleware(['role:admin']);
+Route::get('panel/unionpedidos/{unionpedido}/edit', [UnionpedidoController::class,'edit'])->name('unionpedidos.edit')->middleware(['role:admin']);
+Route::post('panel/unionpedidos', [UnionpedidoController::class,'store'])->name('unionpedidos.store')->middleware(['role:admin']);
+Route::patch('panel/unionpedidos/{unionpedido}', [UnionpedidoController::class,'update'])->name('unionpedidos.update')->middleware(['role:admin']);
+Route::delete('panel/unionpedidos/{unionpedido}/destroy', [UnionpedidoController::class,'destroy'])->name('unionpedidos.destroy')->middleware(['role:admin']);
 //CRUD CARRITOS
 Route::get('panel/carritos', [CarritoController::class,'index'])->name('carritos.index');
 Route::get('panel/carritos/create', [CarritoController::class,'create'])->name('carritos.create');
@@ -116,19 +116,19 @@ Route::post('panel/carritos', [CarritoController::class,'store'])->name('carrito
 Route::patch('panel/carritos/{carrito}', [CarritoController::class,'update'])->name('carritos.update');
 Route::delete('panel/carritos/{carrito}/destroy', [CarritoController::class,'destroy'])->name('carritos.destroy');
 //CRUD TALLAS
-Route::get('panel/tallas', [TallaController::class,'index'])->name('tallas.index');
-Route::get('panel/tallas/create', [TallaController::class,'create'])->name('tallas.create');
-Route::get('panel/tallas/{talla}', [TallaController::class,'show'])->name('tallas.show');
-Route::get('panel/tallas/{talla}/edit', [TallaController::class,'edit'])->name('tallas.edit');
+Route::get('panel/tallas', [TallaController::class,'index'])->name('tallas.index')->middleware(['role:admin']);
+Route::get('panel/tallas/create', [TallaController::class,'create'])->name('tallas.create')->middleware(['role:admin']);
+Route::get('panel/tallas/{talla}', [TallaController::class,'show'])->name('tallas.show')->middleware(['role:admin']);
+Route::get('panel/tallas/{talla}/edit', [TallaController::class,'edit'])->name('tallas.edit')->middleware(['role:admin']);
 Route::post('panel/tallas', [TallaController::class,'store'])->name('tallas.store');
 Route::patch('panel/tallas/{talla}', [TallaController::class,'update'])->name('tallas.update');
 Route::delete('panel/tallas/{talla}/destroy', [TallaController::class,'destroy'])->name('tallas.destroy');
 
 //CRUD USERS
-Route::get('panel/usuarios', [UserController::class,'index'])->name('users.index');
-Route::get('panel/usuarios/create', [UserController::class,'create'])->name('users.create');
-Route::get('panel/usuarios/{usuario}', [UserController::class,'show'])->name('users.show');
-Route::get('panel/usuarios/{usuario}/edit', [UserController::class,'edit'])->name('users.edit');
-Route::post('panel/usuarios', [UserController::class,'store'])->name('users.store');
-Route::patch('panel/usuarios/{usuario}', [UserController::class,'update'])->name('users.update');
-Route::delete('panel/usuarios/{usuario}/destroy', [UserController::class,'destroy'])->name('users.destroy');
+Route::get('panel/usuarios', [UserController::class,'index'])->name('users.index')->middleware(['role:admin']);
+Route::get('panel/usuarios/create', [UserController::class,'create'])->name('users.create')->middleware(['role:admin']);
+Route::get('panel/usuarios/{usuario}', [UserController::class,'show'])->name('users.show')->middleware(['role:admin']);
+Route::get('panel/usuarios/{usuario}/edit', [UserController::class,'edit'])->name('users.edit')->middleware(['role:admin']);
+Route::post('panel/usuarios', [UserController::class,'store'])->name('users.store')->middleware(['role:admin']);
+Route::patch('panel/usuarios/{usuario}', [UserController::class,'update'])->name('users.update')->middleware(['role:admin']);
+Route::delete('panel/usuarios/{usuario}/destroy', [UserController::class,'destroy'])->name('users.destroy')->middleware(['role:admin']);
